+++ src/app/components/Header/Header.css
@@ -1,27 +1,92 @@
 .headerStyles {
     background-color: black;
-    display: flex;
-    padding: 8px 20px;
-    align-items: center;
+    border:3px solid blue;
+    outline:thin ridge red;
+    height:90px;
+    display:flex;
+    align-items:center;
+
+    
 }
 
 .headerStyles .headerImage {
-    height: 45px;
-    border-radius: 8px;
-    box-shadow: 5px 5px 5px black;
-    margin-right: 20px;
+    height:67.5px;
+    width:auto;
+    margin:8px;
+
+    border:4px solid blue;
+    border-radius:2px;    
 }
+
+/*On hover border size increases and changes to green*/
+.headerStyles .headerImage:hover {border:6px solid green;}
+
+
+
 
 .headerStyles .headerLinks {
-    display: flex;
-    list-style-type: none;
-    margin: 0;
-    padding: 0;
-    overflow: hidden;
-    margin-right:auto;
-    color:rgb(64, 61, 253);
+    display:flex;
+    flex-direction:row;
+    color:blue;
+    list-style:none;
+    margin:0;
+    padding:0;
+    gap:20px;
+    font-size:40px;
 }
 
+/*For screen sizes smaller than 600px 
+keep the font-size to 30px*/
+@media screen and (max-width:600px ) {
+    .headerStyles .headerLinks li{
+        font-size:30px;
+    }
+}
+
+.dropbtn
+{
+  text-decoration:none;
+  transition:text-decoration 0.19s ease-in-out; 
+}
+
+.dropbtn:hover
+{
+  text-decoration: underline;
+  text-decoration-color:white;
+}
+
+.dropdown-content {
+  display: none;
+  position: absolute;
+  background-color: black;
+  min-width: 160px;
+  box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
+  z-index: 1;
+  font-weight:bold;
+  
+  
+
+}
+
+.dropdown-content a {
+  color: blue;
+  padding: 12px 16px;
+  text-decoration: none;
+  display: block;
+  text-align: left;
+
+  border:2px solid blue;
+  border-radius: 5px;
+    
+}
+
+.dropdown-content a:hover {background-color: rgb(41, 41, 41);}
+
+.dropdown:hover .dropdown-content {
+  display: block;
+}
+
+/*
 .headerStyles .headerLinks li {
     margin-left: 2cm;
 }
@@ -48,3 +113,5 @@
   font-size: 1.2em;
   color:rgb(64, 61, 253);
 }
+
+*/
