+++ src/app/components/HomeBody/HomeBody.css
@@ -2,7 +2,6 @@
     position: relative;
     margin-top: 50px;
     margin-bottom: 50px;
-
 }
 
 .pulse-hover {
@@ -35,16 +34,32 @@
   display:flex;
   margin-left:auto;
   margin-right:auto;
+  margin-top:5%;
    
 }
 
 .InfoCardStyle div
 {
  width:95%;
- margin-left: 20px;
- margin-right:20px;
-
+ margin-left: auto;
+ margin-right:auto;
  
-
  
+}
+
+@media screen and (max-width:600px ) {
+    .play-btn{
+        width:200px;
+    }
+
+    .InfoCardStyle
+    {
+      flex-direction: column;
+      align-items: center;
+      padding-bottom:20px;
+      
+     
+    }
+
+     
 }
