+++ src/app/components/IllustratorsLogo/IllustratorsLogo.css
@@ -1,0 +1,9 @@
+
+.titleLogoFloat
+{
+     position: relative;
+    margin-top: 50px;
+    margin-left:auto;
+    margin-right:auto;
+    
+}
