+++ src/app/globals.css
@@ -2,11 +2,12 @@
 @import "tailwindcss";
 
 
-html, body {
+html, body, #root {
 
-height:100%;
+min-height: 100vh;
 margin: 0;
 padding:0;
+
  
 }
 
@@ -16,10 +17,23 @@ body
   background-image: url('/backgroundImage.jpg');
   background-size:cover;
   background-position: center center;
-  background-repeat: no repeat;
-  background-attachment: scroll;
 
   cursor: url('/Cursor/Galagaship.png')16 16, auto; 
+}
+
+@media screen and (max-width:600px ) {
+    body
+    {
+      background-size: cover;
+      background-position: top center;
+      background-attachment:fixed;
+    }
+}
+
+.formProperties 
+{
+   background-image: url('/InfoCardBackgroundImage.jpg');
+  
 }
