+++ src/app/login/page.tsx
@@ -2,6 +2,7 @@ import LoginForm from "./login";
 import { _verifySession as getSession } from "@/lib/dal";
 import { redirect } from "next/navigation";
 import Header from "../components/Header/Header";
+import IllustratorsLogo from "../components/IllustratorsLogo/IllustratorsLogo";
 
 export default async function SignUpPage() {
   const session = await getSession();
@@ -13,7 +14,7 @@ export default async function SignUpPage() {
   return (
     <>
       <Header />
-
+      <IllustratorsLogo/> 
       <LoginForm />
     </>
   );
