+++ src/app/sign-up/page.tsx
@@ -2,6 +2,7 @@ import SignupForm from "./sign-up";
 import { _verifySession as getSession } from "@/lib/dal";
 import { redirect } from "next/navigation";
 import Header from "../components/Header/Header.jsx";
+import IllustratorsLogo from "../components/IllustratorsLogo/IllustratorsLogo";
 
 export default async function SignUpPage() {
   const session = await getSession();
@@ -12,7 +13,9 @@ export default async function SignUpPage() {
 
   return (
     <>
+       
       <Header />
+      <IllustratorsLogo/> 
       <SignupForm />
     </>
   );
